requirements.txt
django==4.2
python-decouple
pillow
djangorestframework
markdown 
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated

#testing
pytest
pytest-django